requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
choicelib==0.1.5
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.11
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==2.0.3
vbml==1.1.post1
vk-api==11.9.9
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.2
sqlalchemy==2.0.18
python-dotenv==1.0.0